# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {},
  "paths": {
...
    "/banking.v1.AuthService/IssueUserToken": {
      "post": {
        "tags": [
          "banking.v1.AuthService"
        ],
        "summary": "IssueUserToken",
        "description": "Mint the money-plane token pair for an already-identified user, keyed by their\n concierge user id. Service-to-service: authenticated by the shared issuance token\n in the `authorization` metadata (Bearer), never a user token. The user must already\n exist locally (provisioned by the cross-plane bridge) and not be frozen.",
        "operationId": "banking.v1.AuthService.IssueUserToken",
        "parameters": [
          {
            "name": "Connect-Protocol-Version",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/connect-protocol-version"
            }
          },
          {
            "name": "Connect-Timeout-Ms",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/connect-timeout-header"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/banking.v1.IssueUserTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/connect.error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/banking.v1.TokenResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/banking.v1.AuthService/Refresh": {
      "post": {
        "tags": [
          "banking.v1.AuthService"
        ],
        "summary": "Refresh",
        "description": "Rotate a refresh token: returns a fresh access+refresh pair and invalidates\n the presented refresh token (rotation + reuse detection in the central store).",
        "operationId": "banking.v1.AuthService.Refresh",
        "parameters": [
          {
            "name": "Connect-Protocol-Version",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/connect-protocol-version"
            }
          },
          {
            "name": "Connect-Timeout-Ms",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/connect-timeout-header"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/banking.v1.RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/connect.error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/banking.v1.TokenResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/banking.v1.FundsService/Subscribe": {
      "post": {
        "tags": [
          "banking.v1.FundsService"
        ],
        "summary": "Subscribe",
        "description": "Subscribe `amount` of free balance into a fund, minting units at the current NAV.\n Returns the mint receipt; the holding updates once the relay applies it.",
        "operationId": "banking.v1.FundsService.Subscribe",
        "parameters": [
          {
            "name": "Connect-Protocol-Version",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/connect-protocol-version"
            }
          },
          {
            "name": "Connect-Timeout-Ms",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/connect-timeout-header"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/banking.v1.SubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/connect.error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/banking.v1.Subscription"
                }
              }
            }
          }
        }
      }
    },
...
    "/concierge.v1.AuthService/Exchange": {
      "post": {
        "tags": [
          "concierge.v1.AuthService"
        ],
        "summary": "Exchange",
        "description": "Exchange a Google authorization code (with its PKCE verifier) for concierge's\n OWN tokens. Called by the cabinet BFF on the OAuth callback. Provisions the\n user on first sign-in in the directory module.",
        "operationId": "concierge.v1.AuthService.Exchange",
        "parameters": [
          {
            "name": "Connect-Protocol-Version",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/connect-protocol-version"
            }
          },
          {
            "name": "Connect-Timeout-Ms",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/connect-timeout-header"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/concierge.v1.ExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/connect.error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/concierge.v1.TokenResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/concierge.v1.AuthService/Refresh": {
      "post": {
        "tags": [
          "concierge.v1.AuthService"
        ],
        "summary": "Refresh",
        "description": "Rotate a refresh token: returns a fresh access+refresh pair and invalidates\n the presented refresh token (rotation + reuse detection in the central store).",
        "operationId": "concierge.v1.AuthService.Refresh",
        "parameters": [
          {
            "name": "Connect-Protocol-Version",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/connect-protocol-version"
            }
          },
          {
            "name": "Connect-Timeout-Ms",
            "in": "header",
            "schema": {
              "$ref": "#/components/schemas/connect-timeout-header"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/concierge.v1.RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/connect.error"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/concierge.v1.TokenResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "banking.v1.IssueUserTokenRequest": {
        "type": "object",
        "properties": {
          "concierge_user_id": {
            "type": "string",
            "title": "concierge_user_id",
            "description": "The concierge canonical user id (a UUID) of the already-authenticated user. Banking\n maps it to its local mirror row (provisioned by the cross-plane bridge) and mints the\n money-plane pair for that user. Banking never sees Google's `sub` — the bridge carries\n it; the BFF carries this concierge id, which it has from the concierge sign-in."
          },
          "user_agent": {
            "type": "string",
            "title": "user_agent",
            "description": "Device metadata captured by the BFF at sign-in, stored on the refresh-token family\n for the \"sessions \u0026 devices\" surface. Best-effort, free-form."
          },
          "ip": {
            "type": "string",
            "title": "ip"
          }
        },
        "title": "IssueUserTokenRequest",
        "additionalProperties": false
      },
...
      "banking.v1.RefreshRequest": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "refresh_token"
          }
        },
        "title": "RefreshRequest",
        "additionalProperties": false
      },
...
      "banking.v1.SubscribeRequest": {
        "type": "object",
        "properties": {
          "service": {
            "type": "string",
            "title": "service",
            "description": "the fund/service id"
          },
          "amount": {
            "type": "string",
            "title": "amount",
            "description": "decimal USDT to subscribe (cash in)"
          }
        },
        "title": "SubscribeRequest",
        "additionalProperties": false
      },
      "banking.v1.Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "title": "id"
          },
          "service": {
            "type": "string",
            "title": "service"
          },
          "cash": {
            "type": "string",
            "title": "cash"
          },
          "nav": {
            "type": "string",
            "title": "nav"
          },
          "units": {
            "type": "string",
            "title": "units"
          }
        },
        "title": "Subscription",
        "additionalProperties": false,
        "description": "The receipt for one subscription (mint): `cash` of the user's balance bought `units`\n at `nav`. The position itself is read via GetPosition once the relay has applied it."
      },
      "banking.v1.TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "title": "access_token"
          },
          "access_expires_at": {
            "type": [
              "integer",
              "string"
            ],
            "title": "access_expires_at",
            "format": "int64"
          },
          "refresh_token": {
            "type": "string",
            "title": "refresh_token"
          },
          "refresh_expires_at": {
            "type": [
              "integer",
              "string"
            ],
            "title": "refresh_expires_at",
            "format": "int64"
          },
          "user": {
            "title": "user",
            "$ref": "#/components/schemas/banking.v1.UserSummary"
          }
        },
        "title": "TokenResponse",
        "additionalProperties": false,
        "description": "A first-party token pair plus a snapshot of the principal. All expiries are\n unix SECONDS (matching the JWT `exp` claim); the BFF parses them with\n proto-loader `longs: String`."
      },
...
      "banking.v1.UserSummary": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "title": "user_id",
            "description": "The hub's canonical user id (a UUID) — this is the JWT `sub`, never Google's sub."
          },
          "email": {
            "type": "string",
            "title": "email"
          },
          "status": {
            "type": "string",
            "title": "status"
          },
          "token_version": {
            "type": [
              "integer",
              "string"
            ],
            "title": "token_version",
            "format": "int64"
          },
          "role": {
            "type": "string",
            "title": "role",
            "description": "Kept wire-identical with concierge.v1.UserSummary (the cabinet TS is aliased to the\n concierge shape). The money plane doesn't serve the role here — the cabinet reads it\n from concierge's login summary — so this stays empty; the field exists for parity."
          }
        },
        "title": "UserSummary",
        "additionalProperties": false,
        "description": "Minimal principal snapshot returned with a freshly issued token pair, so the BFF\n can populate a session without a second round trip."
      },
...
      "concierge.v1.ExchangeRequest": {
        "type": "object",
        "properties": {
          "auth_code": {
            "type": "string",
            "title": "auth_code",
            "description": "The Google authorization code from the OAuth redirect."
          },
          "code_verifier": {
            "type": "string",
            "title": "code_verifier",
            "description": "The PKCE code verifier the BFF generated for this transaction (S256)."
          },
          "redirect_uri": {
            "type": "string",
            "title": "redirect_uri",
            "description": "The exact redirect_uri used in the authorize request (Google requires a match)."
          },
          "nonce": {
            "type": "string",
            "title": "nonce",
            "description": "The nonce the BFF put in the authorize request; matched against Google's id_token."
          },
          "user_agent": {
            "type": "string",
            "title": "user_agent",
            "description": "Device metadata captured by the BFF at sign-in, stored on the refresh-token family\n for the \"sessions \u0026 devices\" surface. Best-effort, free-form."
          },
          "ip": {
            "type": "string",
            "title": "ip"
          }
        },
        "title": "ExchangeRequest",
        "additionalProperties": false
      },
...
      "concierge.v1.RefreshRequest": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "refresh_token"
          }
        },
        "title": "RefreshRequest",
        "additionalProperties": false
      },
...
      "concierge.v1.TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "title": "access_token"
          },
          "access_expires_at": {
            "type": [
              "integer",
              "string"
            ],
            "title": "access_expires_at",
            "format": "int64"
          },
          "refresh_token": {
            "type": "string",
            "title": "refresh_token"
          },
          "refresh_expires_at": {
            "type": [
              "integer",
              "string"
            ],
            "title": "refresh_expires_at",
            "format": "int64"
          },
          "user": {
            "title": "user",
            "$ref": "#/components/schemas/concierge.v1.UserSummary"
          }
        },
        "title": "TokenResponse",
        "additionalProperties": false,
        "description": "A first-party token pair plus a snapshot of the principal. All expiries are\n unix SECONDS (matching the JWT `exp` claim); the BFF parses them with\n proto-loader `longs: String`."
      },
...
      "concierge.v1.UserSummary": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "title": "user_id",
            "description": "The concierge canonical user id (a UUID) — this is the JWT `sub`, never Google's sub."
          },
          "email": {
            "type": "string",
            "title": "email"
          },
          "status": {
            "type": "string",
            "title": "status"
          },
          "token_version": {
            "type": [
              "integer",
              "string"
            ],
            "title": "token_version",
            "format": "int64"
          },
          "role": {
            "type": "string",
            "title": "role",
            "description": "The user's platform access role (snake_case: investor/operator/admin/owner), so the\n BFF can gate the admin console at sign-in without a second round trip. Read-only."
          }
        },
        "title": "UserSummary",
        "additionalProperties": false,
        "description": "Minimal principal snapshot returned with a freshly issued token pair, so the BFF\n can populate a session without a second round trip."
      },
      "connect-protocol-version": {
        "type": "number",
        "title": "Connect-Protocol-Version",
        "enum": [
          1
        ],
        "description": "Define the version of the Connect protocol",
        "const": 1
      },
      "connect-timeout-header": {
        "type": "number",
        "title": "Connect-Timeout-Ms",
        "description": "Define the timeout, in ms"
      },
      "connect.error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "examples": [
              "not_found"
            ],
            "enum": [
              "canceled",
              "unknown",
              "invalid_argument",
              "deadline_exceeded",
              "not_found",
              "already_exists",
              "permission_denied",
              "resource_exhausted",
              "failed_precondition",
              "aborted",
              "out_of_range",
              "unimplemented",
              "internal",
              "unavailable",
              "data_loss",
              "unauthenticated"
            ],
            "description": "The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code]."
          },
          "message": {
            "type": "string",
            "description": "A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client."
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/connect.error_details.Any"
            },
            "description": "A list of messages that carry the error details. There is no limit on the number of messages."
          }
        },
        "title": "Connect Error",
        "additionalProperties": true,
        "description": "Error type returned by Connect: https://connectrpc.com/docs/go/errors/#http-representation"
      },
      "connect.error_details.Any": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "A URL that acts as a globally unique identifier for the type of the serialized message. For example: `type.googleapis.com/google.rpc.ErrorInfo`. This is used to determine the schema of the data in the `value` field and is the discriminator for the `debug` field."
          },
          "value": {
            "type": "string",
            "format": "binary",
            "description": "The Protobuf message, serialized as bytes and base64-encoded. The specific message type is identified by the `type` field."
          },
          "debug": {
            "oneOf": [
              {
                "type": "object",
                "title": "Any",
                "additionalProperties": true,
                "description": "Detailed error information."
              }
            ],
            "discriminator": {
              "propertyName": "type"
            },
            "title": "Debug",
            "description": "Deserialized error detail payload. The 'type' field indicates the schema. This field is for easier debugging and should not be relied upon for application logic."
          }
        },
        "additionalProperties": true,
        "description": "Contains an arbitrary serialized message along with a @type that describes the type of the serialized message, with an additional debug field for ConnectRPC error details."
      }
    }
  },
  "security": [],
...
}